seperate task manage
