+++ src/Monai.Deploy.WorkflowManager.sln
@@ -1,21 +1,4 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-
-<!--
-  ~ Copyright 2022 MONAI Consortium
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
--->
-
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.32328.378
 MinimumVisualStudioVersion = 10.0.40219.1
@@ -59,8 +42,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManage
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.TaskManager.Argo", "TaskManager\Plug-ins\Argo\Monai.Deploy.WorkflowManager.TaskManager.Argo.csproj", "{1D5EC18D-96AE-4F67-803E-595E1621C664}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.TaskManager.Runner", "TaskManager\Runner\Monai.Deploy.WorkflowManager.TaskManager.Runner.csproj", "{93129B08-1983-40E4-805E-C2A0BC245839}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.TaskManager.Argo.Tests", "..\tests\UnitTests\TaskManager.Argo.Tests\Monai.Deploy.WorkflowManager.TaskManager.Argo.Tests.csproj", "{2DA40575-4748-4198-BE57-F4AF070DE8E3}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.TaskManager.Tests", "..\tests\UnitTests\TaskManager.Tests\Monai.Deploy.WorkflowManager.TaskManager.Tests.csproj", "{89D3D817-CCFE-4933-9089-D1283F2EA1B5}"
@@ -85,7 +66,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManage
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests", "..\tests\IntegrationTests\TaskManager.IntegrationTests\Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests.csproj", "{A966A7B9-2D4E-4A93-8D20-BD140E7A7F85}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monai.Deploy.WorkflowManager.TaskManager.Database", "TaskManager\Database\Monai.Deploy.WorkflowManager.TaskManager.Database.csproj", "{1ACB9766-0227-430E-ACC2-C2F87AB65FC5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monai.Deploy.WorkflowManager.TaskManager.Database", "TaskManager\Database\Monai.Deploy.WorkflowManager.TaskManager.Database.csproj", "{1ACB9766-0227-430E-ACC2-C2F87AB65FC5}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Plug-ins", "Plug-ins", "{541C5347-5D7D-44B7-95D3-B6FB3D9EB955}"
 EndProject
@@ -167,10 +148,6 @@ Global
 		{1D5EC18D-96AE-4F67-803E-595E1621C664}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D5EC18D-96AE-4F67-803E-595E1621C664}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1D5EC18D-96AE-4F67-803E-595E1621C664}.Release|Any CPU.Build.0 = Release|Any CPU
-		{93129B08-1983-40E4-805E-C2A0BC245839}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{93129B08-1983-40E4-805E-C2A0BC245839}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{93129B08-1983-40E4-805E-C2A0BC245839}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{93129B08-1983-40E4-805E-C2A0BC245839}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2DA40575-4748-4198-BE57-F4AF070DE8E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2DA40575-4748-4198-BE57-F4AF070DE8E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2DA40575-4748-4198-BE57-F4AF070DE8E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -237,7 +214,6 @@ Global
 		{5FF2D787-4A97-4FE0-9533-9E490FCACEA9} = {78B0EB6E-2636-48A7-8FDB-4334C3F3FFF1}
 		{CE4FF14A-6399-44E2-BD29-A2D3F7D1ABB4} = {78B0EB6E-2636-48A7-8FDB-4334C3F3FFF1}
 		{1D5EC18D-96AE-4F67-803E-595E1621C664} = {541C5347-5D7D-44B7-95D3-B6FB3D9EB955}
-		{93129B08-1983-40E4-805E-C2A0BC245839} = {78B0EB6E-2636-48A7-8FDB-4334C3F3FFF1}
 		{2DA40575-4748-4198-BE57-F4AF070DE8E3} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
 		{89D3D817-CCFE-4933-9089-D1283F2EA1B5} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
 		{918E4DE3-A7BF-4B7F-9B5A-5C36FEFA3C30} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
@@ -246,7 +222,7 @@ Global
 		{013395D4-2091-4AB8-96CB-867C0F982096} = {541C5347-5D7D-44B7-95D3-B6FB3D9EB955}
 		{57D00445-0798-4D86-97D4-1586E1B020BC} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
 		{91A0D599-472F-4238-A1A3-07807F9C5F61} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
-		{7D85E95C-A263-429F-BF8B-8F4A922FD579} = {78B0EB6E-2636-48A7-8FDB-4334C3F3FFF1}
+		{7D85E95C-A263-429F-BF8B-8F4A922FD579} = {541C5347-5D7D-44B7-95D3-B6FB3D9EB955}
 		{A966A7B9-2D4E-4A93-8D20-BD140E7A7F85} = {71DDEE7B-E213-4E39-A7F4-4646783A27F7}
 		{1ACB9766-0227-430E-ACC2-C2F87AB65FC5} = {78B0EB6E-2636-48A7-8FDB-4334C3F3FFF1}
 		{541C5347-5D7D-44B7-95D3-B6FB3D9EB955} = {78B0EB6E-2636-48A7-8FDB-4334C3F3FFF1}
