--- src/TaskManager/Runner/Program.cs
@@ -1,230 +1,0 @@
-/*
- * Copyright 2022 MONAI Consortium
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-using System.Globalization;
-using System.IO.Abstractions;
-using Ardalis.GuardClauses;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using Monai.Deploy.Messaging;
-using Monai.Deploy.Messaging.API;
-using Monai.Deploy.Messaging.Configuration;
-using Monai.Deploy.Messaging.Events;
-using Monai.Deploy.Messaging.Messages;
-using Monai.Deploy.Storage;
-using Monai.Deploy.Storage.Configuration;
-using Monai.Deploy.TaskManager.API;
-using Monai.Deploy.WorkflowManager.Common;
-using Monai.Deploy.WorkflowManager.Common.Interfaces;
-using Monai.Deploy.WorkflowManager.Common.Services;
-using Monai.Deploy.WorkflowManager.Configuration;
-using Monai.Deploy.WorkflowManager.Database.Interfaces;
-using Monai.Deploy.WorkflowManager.Database.Options;
-using Monai.Deploy.WorkflowManager.Database.Repositories;
-using Monai.Deploy.WorkflowManager.Storage.Services;
-using Monai.Deploy.WorkflowManager.TaskManager.Argo;
-using Monai.Deploy.WorkflowManager.TaskManager.Argo.StaticValues;
-using Monai.Deploy.WorkflowManager.TaskManager.Database;
-using Monai.Deploy.WorkflowManager.TaskManager.Services;
-using MongoDB.Driver;
-
-namespace Monai.Deploy.WorkflowManager.TaskManager.Runner
-{
-    internal class Program
-    {
-        protected Program()
-        { }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "CA2254:Template should be a static expression", Justification = "Test application")]
-        private static async Task Main(string[] args)
-        {
-            var exitEvent = new ManualResetEvent(false);
-            var host = CreateHostBuilder(args).Build();
-            _ = host.StartAsync();
-
-            var messagingKeys = new MessageBrokerConfigurationKeys();
-            var taskManager = host.Services.GetRequiredService<TaskManager>();
-            Guard.Against.NullService(taskManager, nameof(TaskManager));
-            var logger = host.Services.GetRequiredService<ILogger<Program>>();
-            Guard.Against.NullService(logger, nameof(ILogger<Program>));
-            var publisher = host.Services.GetRequiredService<IMessageBrokerPublisherService>();
-            Guard.Against.NullService(publisher, nameof(IMessageBrokerPublisherService));
-            var subscriber = host.Services.GetRequiredService<IMessageBrokerSubscriberService>();
-            Guard.Against.NullService(subscriber, nameof(IMessageBrokerSubscriberService));
-            var wmConfig = host.Services.GetRequiredService<IOptions<WorkflowManagerOptions>>();
-            Guard.Against.NullService(wmConfig, nameof(IOptions<WorkflowManagerOptions>));
-
-            subscriber.Subscribe(messagingKeys.TaskUpdateRequest, messagingKeys.TaskUpdateRequest, (e) =>
-            {
-                logger.LogInformation($"{e.Message.MessageDescription} received.");
-                var updateMessage = e.Message.ConvertToJsonMessage<TaskUpdateEvent>();
-
-                logger.LogInformation($"Task updated with new status: {updateMessage.Body.Status}");
-                subscriber.Acknowledge(e.Message);
-            }, 1);
-
-            while (taskManager.Status != Contracts.Rest.ServiceStatus.Running)
-            {
-                logger.LogInformation($"Waiting for Task Manager to be ready: state={taskManager.Status}...");
-                await Task.Delay(100).ConfigureAwait(false);
-            }
-            Console.CancelKeyPress += (sender, eventArgs) =>
-            {
-                eventArgs.Cancel = true;
-                exitEvent.Set();
-            };
-
-            // await Task.Run(() =>
-            // {
-            //     var message = GenerateDispatchEvent(args[0], wmConfig.Value);
-            //     logger.LogInformation($"Queuing new job with correlation ID={message.CorrelationId}.");
-            //     publisher.Publish(messagingKeys.TaskDispatchRequest, message);
-            // }).ConfigureAwait(false);
-
-            exitEvent.WaitOne();
-            logger.LogInformation("Stopping Task Manager...");
-
-            await host.StopAsync(CancellationToken.None).ConfigureAwait(false);
-        }
-
-        private static Message GenerateDispatchEvent(string argBaseUri, WorkflowManagerOptions wmConfig)
-        {
-            var correlationId = Guid.NewGuid().ToString();
-            var message = new JsonMessage<TaskDispatchEvent>(new TaskDispatchEvent
-            {
-                WorkflowInstanceId = Guid.NewGuid().ToString(),
-                TaskId = Guid.NewGuid().ToString(),
-                ExecutionId = Guid.NewGuid().ToString(),
-                CorrelationId = correlationId,
-                TaskPluginType = PluginStrings.Argo,
-            }, applicationId: "TaskManagerRunner", correlationId: correlationId, deliveryTag: "1");
-            message.Body.TaskPluginArguments.Add(Keys.BaseUrl, argBaseUri);
-            message.Body.TaskPluginArguments.Add(Keys.WorkflowTemplateName, "list-input-artifacts-template");
-            message.Body.TaskPluginArguments.Add(Keys.MessagingEnddpoint, @$"{wmConfig.Messaging.PublisherSettings["endpoint"]}/{wmConfig.Messaging.PublisherSettings["virtualHost"]}");
-            message.Body.TaskPluginArguments.Add(Keys.MessagingUsername, wmConfig.Messaging.PublisherSettings["username"]);
-            message.Body.TaskPluginArguments.Add(Keys.MessagingPassword, wmConfig.Messaging.PublisherSettings["password"]);
-            message.Body.TaskPluginArguments.Add(Keys.MessagingExchange, wmConfig.Messaging.PublisherSettings["exchange"]);
-            message.Body.TaskPluginArguments.Add(Keys.MessagingTopic, wmConfig.Messaging.Topics.TaskCallbackRequest);
-            message.Body.Inputs.Add(new Messaging.Common.Storage
-            {
-                Name = "input-dicom",
-                Endpoint = wmConfig.Storage.Settings["endpoint"],
-                Bucket = wmConfig.Storage.Settings["bucket"],
-                SecuredConnection = Convert.ToBoolean(wmConfig.Storage.Settings["securedConnection"], CultureInfo.InvariantCulture),
-                RelativeRootPath = "/e08b7d7d-f30c-4f31-87d5-8ce5049aa956/dcm"
-            });
-            message.Body.Inputs.Add(new Messaging.Common.Storage
-            {
-                Name = "input-ehr",
-                Endpoint = wmConfig.Storage.Settings["endpoint"],
-                Bucket = wmConfig.Storage.Settings["bucket"],
-                SecuredConnection = Convert.ToBoolean(wmConfig.Storage.Settings["securedConnection"], CultureInfo.InvariantCulture),
-                RelativeRootPath = "/e08b7d7d-f30c-4f31-87d5-8ce5049aa956/ehr"
-            });
-            message.Body.IntermediateStorage = new Messaging.Common.Storage
-            {
-                Name = "tempStorage",
-                Endpoint = wmConfig.Storage.Settings["endpoint"],
-                Bucket = wmConfig.Storage.Settings["bucket"],
-                SecuredConnection = Convert.ToBoolean(wmConfig.Storage.Settings["securedConnection"], CultureInfo.InvariantCulture),
-                RelativeRootPath = "/rabbit"
-            };
-            message.Body.Outputs.Add(new Messaging.Common.Storage
-            {
-                Name = "tempStorage",
-                Endpoint = wmConfig.Storage.Settings["endpoint"],
-                Bucket = wmConfig.Storage.Settings["bucket"],
-                SecuredConnection = Convert.ToBoolean(wmConfig.Storage.Settings["securedConnection"], CultureInfo.InvariantCulture),
-                RelativeRootPath = "/rabbit"
-            });
-            message.Body.PayloadId = "e08b7d7d-f30c-4f31-87d5-8ce5049aa956";
-            return message.ToMessage();
-        }
-
-        private static void PrintHelp()
-        {
-            Console.WriteLine($"Arguments: Argo_endpoint MinIO_endpoint");
-            Console.WriteLine($"\te.g.: 'http://argo:2746/'");
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureHostConfiguration(configHost =>
-                {
-                    configHost.SetBasePath(Directory.GetCurrentDirectory());
-                    configHost.AddCommandLine(args);
-                })
-                .ConfigureAppConfiguration((builderContext, config) =>
-                {
-                    var env = builderContext.HostingEnvironment;
-                    config
-                        .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-                        .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true, reloadOnChange: true)
-                        .AddEnvironmentVariables();
-                })
-                .ConfigureLogging((builderContext, configureLogging) =>
-                {
-                    configureLogging.AddConfiguration(builderContext.Configuration.GetSection("Logging"));
-                    configureLogging.AddFile(o => o.RootPath = AppContext.BaseDirectory);
-                })
-                .ConfigureServices((hostContext, services) =>
-                {
-                    services.AddOptions<WorkflowManagerOptions>().Bind(hostContext.Configuration.GetSection("WorkflowManager"));
-                    services.AddOptions<StorageServiceConfiguration>().Bind(hostContext.Configuration.GetSection("WorkflowManager:storage"));
-                    services.AddOptions<MessageBrokerServiceConfiguration>().Bind(hostContext.Configuration.GetSection("WorkflowManager:messaging"));
-                    services.AddHttpClient();
-
-                    services.AddMonaiDeployStorageService(hostContext.Configuration.GetSection("WorkflowManager:storage:serviceAssemblyName").Value);
-                    services.AddMonaiDeployMessageBrokerPublisherService(hostContext.Configuration.GetSection("WorkflowManager:messaging:publisherServiceAssemblyName").Value);
-                    services.AddMonaiDeployMessageBrokerSubscriberService(hostContext.Configuration.GetSection("WorkflowManager:messaging:subscriberServiceAssemblyName").Value);
-
-                    // Mongo DB (Workflow Manager)
-                    services.Configure<WorkloadManagerDatabaseSettings>(hostContext.Configuration.GetSection("WorkloadManagerDatabase"));
-                    services.Configure<TaskManagerDatabaseSettings>(hostContext.Configuration.GetSection("WorkloadManagerDatabase"));
-                    services.AddSingleton<IMongoClient, MongoClient>(s => new MongoClient(hostContext.Configuration["WorkloadManagerDatabase:ConnectionString"]));
-                    services.AddTransient<ITaskDispatchEventRepository, TaskDispatchEventRepository>();
-                    services.AddTransient<IWorkflowRepository, WorkflowRepository>();
-                    services.AddTransient<IWorkflowInstanceRepository, WorkflowInstanceRepository>();
-                    services.AddTransient<IPayloadRepsitory, PayloadRepository>();
-                    services.AddTransient<IWorkflowService, WorkflowService>();
-                    services.AddTransient<IWorkflowInstanceService, WorkflowInstanceService>();
-                    services.AddTransient<IPayloadService, PayloadService>();
-                    services.AddTransient<IDicomService, DicomService>();
-
-                    services.AddSingleton<TaskManager>();
-                    services.AddSingleton<IArgoProvider, ArgoProvider>();
-                    services.AddSingleton<IKubernetesProvider, KubernetesProvider>();
-                    services.AddTransient<IFileSystem, FileSystem>();
-                    services.AddTransient<ITaskDispatchEventService, TaskDispatchEventService>();
-
-                    services.AddHttpClient("Argo");
-                    services.AddHttpClient("Argo-Insecure").ConfigurePrimaryHttpMessageHandler(() => new HttpClientHandler
-                    {
-                        ClientCertificateOptions = ClientCertificateOption.Manual,
-                        ServerCertificateCustomValidationCallback =
-                            (httpRequestMessage, cert, cetChain, policyErrors) =>
-                            {
-                                return true;
-                            }
-                    });
-
-                    services.AddHostedService<TaskManager>(p => p.GetRequiredService<TaskManager>());
-                });
-    }
-}
