+++ src/TaskManager/appsettings.Development.json
@@ -1,14 +1,17 @@
 {
   "WorkloadManagerDatabase": {
-    "ConnectionString": "mongodb://root:rootpassword@127.0.0.1:27017",
-    "DatabaseName": "TaskManager",
+    "ConnectionString": "mongodb://root:rootpassword@localhost:27017",
+    "DatabaseName": "WorkloadManager",
     "WorkflowCollectionName": "Workflows",
     "WorkflowInstanceCollectionName": "WorkflowInstances",
-    "PayloadCollectionName": "Payloads",
-    "TaskDispatchEventCollectionName": "TaskDispatchEvents"
+    "PayloadCollectionName": "Payloads"
   },
   "WorkflowManager": {
-    "taskManager": {
+    "endpointSettings": {
+      "defaultPageSize": 10,
+      "maxPageSize": 1000000
+    },
+    "TaskManager": {
       "concurrency": 1,
       "plug-ins": {
         "argo": "Monai.Deploy.WorkflowManager.TaskManager.Argo.ArgoPlugin, Monai.Deploy.WorkflowManager.TaskManager.Argo",
@@ -20,39 +23,51 @@
         "test": "Monai.Deploy.WorkflowManager.TaskManager.TestPlugin.Repositories.TestPluginRepository, Monai.Deploy.WorkflowManager.TaskManager.TestPlugin"
       }
     },
-    "messaging": {
-      "publisherServiceAssemblyName": "Monai.Deploy.Messaging.RabbitMQ.RabbitMQMessagePublisherService, Monai.Deploy.Messaging.RabbitMQ",
-      "publisherSettings": {
-        "endpoint": "10.110.54.121",
-        "username": "mdig",
-        "password": "gidm",
-        "virtualHost": "monaideploy",
-        "exchange": "monaideploy"
-      },
-      "subscriberServiceAssemblyName": "Monai.Deploy.Messaging.RabbitMQ.RabbitMQMessageSubscriberService, Monai.Deploy.Messaging.RabbitMQ",
-      "subscriberSettings": {
-        "endpoint": "10.110.54.121",
-        "username": "mdig",
-        "password": "gidm",
-        "virtualHost": "monaideploy",
-        "exchange": "monaideploy",
-        "exportRequestQueue": "export_tasks",
-        "deadLetterExchange": "monaideploy-dead-letter",
-        "deliveryLimit": 3,
-        "requeueDelay": 30
-      }
-    },
     "storage": {
       "serviceAssemblyName": "Monai.Deploy.Storage.MinIO.MinIoStorageService, Monai.Deploy.Storage.MinIO",
       "settings": {
-        "bucket": "monaideploy",
-        "endpoint": "10.105.3.254",
-        "accessKey": "monaideploy",
-        "accessToken": "monaideploy",
-        "securedConnection": false,
+        "endpoint": "localhost:9000",
+        "accessKey": "minioadmin",
+        "accessToken": "minioadmin",
+        "bucket": "test-bucket",
         "region": "eu-west-2",
-        "executableLocation": "/usr/bin/mc",
-        "serviceName": "test"
+        "securedConnection": false,
+        "executableLocation": "/.",
+        "serviceName": "serviceName"
+      }
+    },
+    "messaging": {
+      "publisherServiceAssemblyName": "Monai.Deploy.Messaging.RabbitMQ.RabbitMQMessagePublisherService, Monai.Deploy.Messaging.RabbitMQ",
+      "publisherSettings": {
+        "endpoint": "localhost",
+        "username": "admin",
+        "password": "admin",
+        "virtualHost": "monaideploy",
+        "exchange": "monaideploy",
+        "deadLetterExchange": "monaideploy-dead-letter",
+        "deliveryLimit": 3,
+        "requeueDelay": 30
+      },
+      "subscriberServiceAssemblyName": "Monai.Deploy.Messaging.RabbitMQ.RabbitMQMessageSubscriberService, Monai.Deploy.Messaging.RabbitMQ",
+      "subscriberSettings": {
+        "endpoint": "localhost",
+        "username": "admin",
+        "password": "admin",
+        "virtualHost": "monaideploy",
+        "exchange": "monaideploy",
+        "exportRequestQueue": "export_tasks"
+      },
+      "retries": {
+        "retryDelays": "localhost"
+      },
+      "topics": {
+        "workflowRequest": "md.workflow.request",
+        "exportComplete": "md.export.complete",
+        "exportRequestPrefix": "md.export.request"
+      },
+      "dicomAgents": {
+        "dicomWebAgentName": "monaidicomweb",
+        "scuAgentName": "monaiscu"
       }
     }
   },
@@ -64,9 +79,10 @@
       "Microsoft.EntityFrameworkCore": "Warning",
       "Microsoft.Hosting.Lifetime": "Information",
       "Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker": "Error",
-      "Monai": "Debug"
+      "Monai": "Information"
     },
     "Console": {
+      "FormatterName": "Systemd",
       "FormatterOptions": {
         "ColorBehavior": "Disabled",
         "IncludeScopes": true,
@@ -74,6 +90,21 @@
         "TimestampFormat": " HH:mm:ss ",
         "UseUtcTimestamp": true
       }
+    },
+    "File": {
+      "BasePath": "Logs",
+      "FileEncodingName": "utf-8",
+      "DateFormat": "yyyyMMdd",
+      "CounterFormat": "000",
+      "MaxFileSize": 10485760,
+      "IncludeScopes": true,
+      "MaxQueueSize": 100,
+      "TextBuilderType": "Monai.Deploy.WorkflowManager.Logging.FileLoggingTextFormatter, Monai.Deploy.WorkflowManager",
+      "Files": [
+        {
+          "Path": "MWM-<date>-<counter>.log"
+        }
+      ]
     }
   },
   "Kestrel": {
