+++ src/TaskManager/appsettings.Local.json
@@ -1,0 +1,120 @@
+{
+  "WorkloadManagerDatabase": {
+    "ConnectionString": "mongodb://root:rootpassword@localhost:30017",
+    "DatabaseName": "WorkloadManager",
+    "WorkflowCollectionName": "Workflows",
+    "WorkflowInstanceCollectionName": "WorkflowInstances",
+    "PayloadCollectionName": "Payloads"
+  },
+  "WorkflowManager": {
+    "endpointSettings": {
+      "defaultPageSize": 10,
+      "maxPageSize": 1000000
+    },
+    "TaskManager": {
+      "concurrency": 1,
+      "plug-ins": {
+        "argo": "Monai.Deploy.WorkflowManager.TaskManager.Argo.ArgoPlugin, Monai.Deploy.WorkflowManager.TaskManager.Argo",
+        "aide_clinical_review": "Monai.Deploy.WorkflowManager.TaskManager.AideClinicalReview.AideClinicalReviewPlugin, Monai.Deploy.WorkflowManager.TaskManager.AideClinicalReview",
+        "test": "Monai.Deploy.WorkflowManager.TaskManager.TestPlugin.TestPlugin, Monai.Deploy.WorkflowManager.TaskManager.TestPlugin"
+      },
+      "meta-data": {
+        "argo": "Monai.Deploy.WorkflowManager.TaskManager.Argo.Repositories.ArgoMetadataRepository, Monai.Deploy.WorkflowManager.TaskManager.Argo",
+        "test": "Monai.Deploy.WorkflowManager.TaskManager.TestPlugin.Repositories.TestPluginRepository, Monai.Deploy.WorkflowManager.TaskManager.TestPlugin"
+      }
+    },
+    "storage": {
+      "serviceAssemblyName": "Monai.Deploy.Storage.MinIO.MinIoStorageService, Monai.Deploy.Storage.MinIO",
+      "settings": {
+        "endpoint": "minio:9000", // note edit C:\Windows\System32\drivers\etc\Hosts and add  `127.0.0.1	minio`
+        "accessKey": "admin",
+        "accessToken": "password",
+        "bucket": "test-bucket",
+        "region": "eu-west-2",
+        "securedConnection": false,
+        "executableLocation": "mc.exe",
+        "serviceName": "serviceName"
+      }
+    },
+    "messaging": {
+      "publisherServiceAssemblyName": "Monai.Deploy.Messaging.RabbitMQ.RabbitMQMessagePublisherService, Monai.Deploy.Messaging.RabbitMQ",
+      "publisherSettings": {
+        "endpoint": "localhost",
+        "username": "admin",
+        "password": "admin",
+        "port": "30072",
+        "virtualHost": "monaideploy",
+        "exchange": "monaideploy",
+        "deadLetterExchange": "monaideploy-dead-letter",
+        "deliveryLimit": 3,
+        "requeueDelay": 30
+      },
+      "subscriberServiceAssemblyName": "Monai.Deploy.Messaging.RabbitMQ.RabbitMQMessageSubscriberService, Monai.Deploy.Messaging.RabbitMQ",
+      "subscriberSettings": {
+        "endpoint": "localhost",
+        "username": "admin",
+        "password": "admin",
+        "port": "30072",
+        "virtualHost": "monaideploy",
+        "exchange": "monaideploy",
+        "exportRequestQueue": "export_tasks"
+      },
+      "retries": {
+        "retryDelays": "localhost"
+      },
+      "topics": {
+        "workflowRequest": "md.workflow.request",
+        "exportComplete": "md.export.complete",
+        "exportRequestPrefix": "md.export.request"
+      },
+      "dicomAgents": {
+        "dicomWebAgentName": "monaidicomweb",
+        "scuAgentName": "monaiscu"
+      }
+    }
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "System": "Warning",
+      "Microsoft": "Warning",
+      "Microsoft.EntityFrameworkCore": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information",
+      "Microsoft.AspNetCore.Mvc.Infrastructure.ControllerActionInvoker": "Error",
+      "Monai": "Information"
+    },
+    "Console": {
+      "FormatterName": "Systemd",
+      "FormatterOptions": {
+        "ColorBehavior": "Disabled",
+        "IncludeScopes": true,
+        "SingleLine": false,
+        "TimestampFormat": " HH:mm:ss ",
+        "UseUtcTimestamp": true
+      }
+    },
+    "File": {
+      "BasePath": "Logs",
+      "FileEncodingName": "utf-8",
+      "DateFormat": "yyyyMMdd",
+      "CounterFormat": "000",
+      "MaxFileSize": 10485760,
+      "IncludeScopes": true,
+      "MaxQueueSize": 100,
+      "TextBuilderType": "Monai.Deploy.WorkflowManager.Logging.FileLoggingTextFormatter, Monai.Deploy.WorkflowManager",
+      "Files": [
+        {
+          "Path": "MWM-<date>-<counter>.log"
+        }
+      ]
+    }
+  },
+  "Kestrel": {
+    "EndPoints": {
+      "Http": {
+        "Url": "http://localhost:5000"
+      }
+    }
+  },
+  "AllowedHosts": "*"
+}
